% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{as_igraph}
\alias{as_igraph}
\title{Convert ICON objects to igraph objects}
\usage{
as_igraph(x, directed = FALSE, weighted = (ncol(x) > 2))
}
\arguments{
\item{x}{edgelist of classes `ICON` and `data.frame` to be coerced}

\item{directed}{`TRUE` if network has directed edges; `FALSE` otherwise}

\item{weighted}{`TRUE` if network has weighted edges; `FALSE` otherwise}
}
\value{
an `igraph` object
}
\description{
`ICON::get_data` returns data frames that also have class `ICON`. However,
conducting network analysis could require using another R package and its
associated network structure and class. The `as_igraph` function converts
`ICON` objects to `igraph` objects. It allows `ICON` users to take full
advantage of the functionality provided in the `igraph` package.
}
\details{
Note that if the `ICON` object has more than two columns, `as_igraph`
assumes that the third column is numeric and represents edge weights.
Use `weighted = FALSE` to ignore existing edge weights and re-order
columns if you would prefer another column be used for edge weights.
}
\examples{
\dontrun{
# download dataset using ICON
get_data("chess")

# don't care about relabeled vertices
converted <- as_igraph(chess, directed = TRUE, weighted = TRUE)

# look at edges
igraph::E(converted)

# look at edge weights
igraph::E(converted)$weight
}
}
