% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICON-package.R
\docType{package}
\name{ICON-package}
\alias{ICON-package}
\title{ICON: Isoform-Cell-type Classification via Overlapping Networks}
\description{
Hierarchical Random Forest pipeline for cell type classification from single-cell RNA-seq data. ICON uses both isoform and gene count matrices and combines predictions from two levels of Random Forest models to improve cell type classification accuracy.
}
\author{
\strong{Maintainer}: Sachini Wijewardena \email{jithma.wijewardena@gmail.com}

Authors:
\itemize{
  \item Sachini Wijewardena \email{jithma.wijewardena@gmail.com}
}

}
\keyword{internal}
