% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{import_count_matrix}
\alias{import_count_matrix}
\title{Import Seurat Object}
\usage{
import_count_matrix(file_path)
}
\arguments{
\item{file_path}{A Seurat object.}
}
\value{
A Seurat object.
}
\description{
Validates that the input is a Seurat object and prints basic summary information.
}
