% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{internet_routers}
\alias{internet_routers}
\title{Internet Routers (1993)}
\format{A data frame with 2 columns and 48436 rows:
\describe{
  \item{Node1}{one of the nodes defining the current edge}
  \item{Node2}{other node defining the current edge}
}}
\source{
\url{https://github.com/gephi/gephi/wiki/Datasets}
}
\usage{
internet_routers
}
\description{
Undirected, unweighted network that is a symmetrized snapshot of the
structure of the Internet at the level of autonomous systems, reconstructed
from BGP tables posted by the University of Oregon Route Views Project at
<http://routeviews.org>. This was created by Mark Newman in 2006.
Network is formatted as an edgelist. Each node is
also associated with numerical labels - these have been omitted from the
dataset loaded by this R pacakge. However, this information can be found
in the raw dataset at this package's GitHub repo at
<https://github.com/rrrlw/ICON> under the "data-raw" directory.
}
\details{
On the webpage at the source URL (below), download the linked GML file
under the "Social Networks" heading associated with the bullet point
titled "Les Miserables".

Citation:
Newman M. Internet (2006). University of Oregon Route Views Project.
<http://routeviews.org>
}
\keyword{datasets}
