% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{network_coauthors}
\alias{network_coauthors}
\title{Coauthorships in Network Science (2006)}
\format{A data frame with 3 columns and 2742 rows:
\describe{
  \item{Node1}{one of the nodes defining the current edge}
  \item{Node2}{other node defining the current edge}
  \item{Weight}{edge weight}
}}
\source{
\url{https://github.com/gephi/gephi/wiki/Datasets}
}
\usage{
network_coauthors
}
\description{
Undirected, weighted coauthorship network of scientists working on network
theory, compiled by M. Newman in May 2006. Network is formatted as an
edgelist. See the cited journal article below to understand how edge weight
was calculated. Each node is also associated with researcher names - these
have been omitted from the dataset loaded by this R pacakge. However,
this information can be found in the raw dataset at this package's GitHub
repo at <https://github.com/rrrlw/ICON> under the "data-raw" directory.
}
\details{
On the webpage at the source URL (below), download the linked GML file
under the "Social Networks" heading associated with the bullet point
titled "Coauthorships in network science".

Citation:
Newman MEJ. Finding community structure in networks using the eigenvectors
of matrices. Phys Rev E. 2006; 74: 036104.
}
\keyword{datasets}
