% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{terrorist_pairs}
\alias{terrorist_pairs}
\title{Terrorist network (2002)}
\format{A data frame with 2 columns and 304 rows:
\describe{
  \item{Node1}{one of the nodes defining the current edge}
  \item{Node2}{other node defining the current edge}
}}
\source{
\url{http://tuvalu.santafe.edu/~aaronc/datacode.htm}
}
\usage{
terrorist_pairs
}
\description{
Undirected, unweighted adjacency network of terrorists.
Network is formatted as an edgelist. Node labels (words) were
provided in the raw dataset, which can be found at the ICON database.
}
\details{
On the webpage at the source URL (below), download the linked folder
under the "Terrorist associations" heading (4th in list).

Citation:
Krebs V. Mapping networks of terrorist cells. Connections. 2002; 24: 43-52.
}
\keyword{datasets}
