% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICON.R
\docType{package}
\name{ICON}
\alias{ICON}
\title{ICON: Provides Easy Access to Complex Systems Datasets}
\description{
Provides easy-to-use and easy-to-access datasets from the
Index of COmplex Networks (ICON) database available on the University
of Colorado website. All datasets can be loaded with a single function
call and new datasets are being slowly added from ICON.
}
